displays data of food vendor plus its menu
